server/requirements.txt
# PDF to Office conversion libraries
# Using newer pdf2docx to fix 'Rect' object bug
pdf2docx==0.5.9
tabula-py==2.9.0
openpyxl==3.1.2
python-pptx==0.6.23
pdf2image==1.17.0
Pillow==10.3.0
# Alternative PDF library for better compatibility
PyMuPDF==1.24.0
# For Word document creation in fallback
python-docx==1.1.0
